Fix sub-second durations, normalise line endings
duration_seconds was an Integer column, so Postgres truncated a 0.42s
analysis to 0 while the in-memory store kept it. The two store
implementations disagreed about the same report, and the UI reported
every fast analysis as taking 0.0s.

Alembic 0003 widens the column to double precision. The cast is lossless,
so existing whole-second values survive exactly; only the precision that
was never written is unrecoverable.

Adds a regression test covering both read paths. PostgresReportStore.list
already coerced the column with float(), which looked like a guard but
was not one — the truncation had happened on write, long before the read.

.gitattributes normalises line endings to LF in the repository. Without
it a checkout on Windows commits CRLF back and every file reads as
wholly rewritten.

+++ app/store.py
@@ -13,7 +13,7 @@ from collections import OrderedDict
 from datetime import datetime, timezone
 from typing import Protocol
 
-from sqlalchemy import DateTime, Integer, String, Text, select
+from sqlalchemy import DateTime, Float, Integer, String, Text, select
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column
 from sqlalchemy.types import JSON
@@ -44,7 +44,10 @@ class ReportRow(Base):
     score: Mapped[int] = mapped_column(Integer, nullable=False)
     grade: Mapped[str] = mapped_column(String(2), nullable=False)
     total_findings: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    duration_seconds: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
+    # Float, not Integer. Stored as an integer, a sub-second analysis rounded to
+    # 0 on Postgres while the in-memory store kept 0.4 — the two implementations
+    # disagreed about the same report, and the UI showed "analysed in 0.0s".
+    duration_seconds: Mapped[float] = mapped_column(Float, nullable=False, default=0.0)
     # Null for anonymous reports, including every row created before sign-in
     # existed. Those stay reachable by id and are never listed.
     owner_id: Mapped[str | None] = mapped_column(String(24), index=True)
@@ -167,7 +170,7 @@ class PostgresReportStore:
             score=report.score.value,
             grade=report.score.grade,
             total_findings=len(report.findings),
-            duration_seconds=int(report.duration_seconds),
+            duration_seconds=report.duration_seconds,
             payload=report.model_dump(mode="json"),
         )
         async with maker() as session:
